+++ app/Http/Controllers/PositionController.php
@@ -4,11 +4,17 @@ namespace App\Http\Controllers;
 
 use App\Models\Position;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Auth; 
 
 class PositionController extends Controller
 {
+    
     public function index()
 {
+    if (!Auth::check() || Auth::id() !== 1) {
+        return redirect()->route('login');
+    }
+
     // Obtém todas as posições do banco de dados
     $positions = Position::all();
 
@@ -18,11 +24,17 @@ class PositionController extends Controller
 
     public function create()
     {
+        if (!Auth::check() || Auth::id() !== 1) {
+            return redirect()->route('login');
+        }
         return view('positions.create');
     }
 
     public function store(Request $request)
     {
+        if (!Auth::check() || Auth::id() !== 1) {
+            return redirect()->route('login');
+        }
         $request->validate([
             'title' => 'required|string|max:100',
             'description' => 'nullable|string|max:255', 
@@ -34,16 +46,28 @@ class PositionController extends Controller
 
     public function show(Position $position)
     {
+        if (!Auth::check() || Auth::id() !== 1) {
+            return redirect()->route('login');
+        }
+
         return view('positions.show', compact('position'));
     }
 
     public function edit(Position $position)
     {
+        if (!Auth::check() || Auth::id() !== 1) {
+            return redirect()->route('login');
+        }
+
         return view('positions.edit', compact('position'));
     }
 
     public function update(Request $request, Position $position)
     {
+        if (!Auth::check() || Auth::id() !== 1) {
+            return redirect()->route('login');
+        }
+
         $request->validate([
             'title' => 'required|string|max:100',
             'description' => 'nullable|string|max:255', 
@@ -55,6 +79,10 @@ class PositionController extends Controller
 
     public function destroy(Position $position)
     {
+        if (!Auth::check() || Auth::id() !== 1) {
+            return redirect()->route('login');
+        }
+        
         $position->delete();
         return redirect()->route('positions.index')->with('success', 'Posição deletada com sucesso.');
     }
